[refactor] Move WongojiComponent to parent package
